+++ tracking/trackFindingCDC/eventdata/trajectories/tests/CDCTrajectory2D.cc
@@ -23,7 +23,7 @@ TEST(TrackFindingCDCTest, eventdata_trajectories_CDCTrajectory2D_constructorPosM
   Vector2D newMom2D(1.0, 2.0);
   Vector2D newPos2D(1.0, 2.0);
   SignType newChargeSign = PLUS;
-  FloatType bZ = 2.0;
+  double bZ = 2.0;
 
   CDCTrajectory2D trajectory(newPos2D, newMom2D, newChargeSign, bZ);
